+++ lib/widget/align.dart
@@ -4,13 +4,13 @@ extension AlignWidgetEx on Widget {
   Align align({
     Key? key,
     Alignment at = Alignment.center,
-    double? widthFactor,
-    double? heightFactor,
+    double? wFactor,
+    double? hFactor,
   }) =>
       Align(
         alignment: at,
-        widthFactor: widthFactor,
-        heightFactor: heightFactor,
+        widthFactor: wFactor,
+        heightFactor: hFactor,
         child: this,
       );
 }